requirements.txt
beautifulsoup4==4.13.3
requests==2.32.3
python-decouple==3.8
python-telegram-bot==21.10